requirements.txt
openai==1.91.0
python-dotenv>=1.0.0
streamlit==1.46.0
